refactor: adjusting the project structure
