+++ client/src/@types/vite-env.d.ts
