+++ client/src/core/App.css
@@ -1,3 +1,7 @@
+body {
+  margin: 0;
+}
+
 #root {
   margin: 0;
   padding: 0;
